requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.8.2
blis==0.4.1
boto==2.49.0
boto3==1.12.16
botocore==1.15.16
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.3
django-chartjs==2.0.0
django-crispy-forms==1.9.0
django-extensions==2.2.8
django-mathfilters==1.0.0
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
future==0.18.2
gast==0.2.2
gensim==3.8.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
hurst==0.0.5
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.5.0
incremental==17.5.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
npm==0.1.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
optional-django==0.1.0
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.6.0
patsy==0.5.1
pdfminer.six==20200124
pdftotree==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
plotly==4.5.1
preshed==3.0.2
prompt-toolkit==3.0.3
Protego==0.1.16
protobuf==3.11.2
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.6
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.11.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.0
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rest-pandas==1.1.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
Scrapy==2.0.1
seaborn==0.10.0
selectivesearch==0.4
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
smart-open==1.9.0
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.3
spacy-lookups-data==0.2.0
sqlparse==0.3.0
srsly==1.0.2
statsmodels==0.11.1
striprtf==0.0.6
tabula-py==2.0.4
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
w3lib==1.21.0
Wand==0.5.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.11.2
xlrd==1.2.0
zipp==3.1.0
zope.interface==5.0.0
